split prompts to separate class
