+++ src/lib/ui/cartAbandonmentConfirmer.js
@@ -1,9 +1,9 @@
 export default class CartAbandonmentConfirmer {
-    constructor(document, rimiDOM, sa2) {
+    constructor(document, rimiDOM, promptService) {
+        this.userAcceptedAbandonment = false;
         this.document = document;
         this.rimiDOM = rimiDOM;
-        this.sa2 = sa2;
-        this.userAcceptedAbandonment = false;
+        this.promptService = promptService;
     }
 
     bindToCartChangeButtons() {
@@ -44,22 +44,17 @@ export default class CartAbandonmentConfirmer {
         }
     }
 
-    askForAbandonmentConfirmation(clickTarget) {
-        this.sa2.fire({
-            text: 'Are you sure you want to abandon your current basket?',
-            showCancelButton: true,
-            confirmButtonText: `Yes`,
-            customClass: {
-                container: 'smart-basket-confirm-action',
-                confirmButton: 'smart-basket-accept',
-                cancelButton: 'smart-basket-cancel'
-            }
-        }).then((result) => {
-            if (result.isConfirmed) {
-                this.userAcceptedAbandonment = true;
-                clickTarget.click();
+    askForAbandonmentConfirmation(clickedCartElement) {
+        this.promptService.promptCartAbandonment().then((result) => {
+            if (result) {
+                this.abandonCart(clickedCartElement);
             }
         })
+    }
+
+    abandonCart(clickedCartElement) {
+        this.userAcceptedAbandonment = true;
+        clickedCartElement.click();
     }
 
     stopEventExecution(event) {
